연구 그룹 state 사용
