+++ app/[locale]/research/groups/page.tsx
@@ -2,52 +2,32 @@ export const dynamic = 'force-static';
 
 import { getResearchGroups } from '@/apis/research';
 
-import SelectionList from '@/components/common/selection/SelectionList';
-import PageLayout from '@/components/layout/pageLayout/PageLayout';
-import ResearchGroupDetails from '@/components/research/groups/ResearchGroupDetails';
+// 추후에 api 연결하면 주석 해제하고 원상복구
 
-import { researchGroups } from '@/types/page';
-import { ResearchGroup } from '@/types/research';
+// import SelectionList from '@/components/common/selection/SelectionList';
+// import PageLayout from '@/components/layout/pageLayout/PageLayout';
+// import ResearchGroupDetails from '@/components/research/groups/ResearchGroupDetails';
+import ResearchGroupsPageContent from '@/components/research/groups/ResearchGroupsPageContent';
 
-import { findSelectedItem } from '@/utils/findSelectedItem';
-import { getPath } from '@/utils/page';
+// import { researchGroups } from '@/types/page';
+// import { ResearchGroup } from '@/types/research';
+
+// import { findSelectedItem } from '@/utils/findSelectedItem';
+// import { getPath } from '@/utils/page';
 
 interface ResearchGroupsPageProps {
   searchParams: { selected?: string };
 }
 
-const researchGroupsPath = getPath(researchGroups);
+// const researchGroupsPath = getPath(researchGroups);
 
 export default async function ResearchGroupsPage({ searchParams }: ResearchGroupsPageProps) {
   const { description, groups } = await getResearchGroups();
-  const selectedGroup = findSelectedItem<ResearchGroup>(
-    groups,
-    decodeURI(searchParams.selected ?? ''),
-    groups[0]?.name,
-  );
-
-  return (
-    <PageLayout titleType="big" titleMargin="mb-9">
-      <ResearchDescription description={description} />
-      <SelectionList
-        names={groups.map((group) => group.name)}
-        selectedItemName={selectedGroup?.name ?? ''}
-        path={researchGroupsPath}
-        listGridColumnClass="grid-cols-[13.25rem_8.75rem_13.75rem_14.5rem]"
-      />
-      {selectedGroup ? (
-        <ResearchGroupDetails group={selectedGroup} />
-      ) : (
-        <p>
-          <b>{`'${searchParams.selected}'`}</b> 연구그룹은 존재하지 않습니다.
-        </p>
-      )}
-    </PageLayout>
-  );
+  return <ResearchGroupsPageContent description={description} groups={groups} />;
 }
 
-function ResearchDescription({ description }: { description: string }) {
-  return (
-    <p className="mb-[3.25rem] font-noto text-sm tracking-wide leading-[1.625rem]">{description}</p>
-  );
-}
+// function ResearchDescription({ description }: { description: string }) {
+//   return (
+//     <p className="mb-[3.25rem] font-noto text-sm tracking-wide leading-[1.625rem]">{description}</p>
+//   );
+// }
